requirements.txt
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
kafka-python
shodan
